feat(NTS-117)!: bind cumulative counters as int, not BigInt
The eight cumulative counters on NtsDnsPoolStats and NtsTrustStatus
were BigInt only because the bridge-facing Rust structs declared them
u64, which FRB binds to BigInt. The wraparound rationale on the fields
justifies the 64-bit backing store, not the BigInt binding.

Widen the bridge-facing fields to i64 so FRB binds PlatformInt64, and
narrow to a plain int in the conversion layer. The backing counters
stay AtomicU64; the wire layout is unchanged at 8 bytes per counter.

This matches what PhaseTimings and ntsBoottimeMicros already did, and
removes the split inside NtsDnsPoolStats, whose inFlight and
highWaterMark were already plain int. Both DTOs are now
const-constructible with literal counters.

Breaking: retargets the in-flight CHANGELOG section to 9.0.

+++ lib/src/api/nts_convert.dart
@@ -78,9 +78,9 @@ PhaseTimings _publicPhase(ffi.PhaseTimings p) => PhaseTimings(
 NtsDnsPoolStats _publicStats(ffi.NtsDnsPoolStats s) => NtsDnsPoolStats(
   inFlight: s.inFlight,
   highWaterMark: s.highWaterMark,
-  recovered: s.recovered,
-  refused: s.refused,
-  spawnFailed: s.spawnFailed,
+  recovered: s.recovered.toInt(),
+  refused: s.refused.toInt(),
+  spawnFailed: s.spawnFailed.toInt(),
 );
 
 TimeoutPhase _publicTimeoutPhase(ffi.TimeoutPhase phase) => switch (phase) {
@@ -187,10 +187,10 @@ NtsTrustStatus _publicTrustStatus(ffi.NtsTrustStatus s) => NtsTrustStatus(
   defaultClientBackend: s.defaultClientBackend == null
       ? null
       : _publicTrustBackend(s.defaultClientBackend!),
-  defaultBackendPlatformCount: s.defaultBackendPlatformCount,
-  defaultBackendHybridCount: s.defaultBackendHybridCount,
-  defaultBackendWebpkiCount: s.defaultBackendWebpkiCount,
-  defaultBackendCustomCount: s.defaultBackendCustomCount,
+  defaultBackendPlatformCount: s.defaultBackendPlatformCount.toInt(),
+  defaultBackendHybridCount: s.defaultBackendHybridCount.toInt(),
+  defaultBackendWebpkiCount: s.defaultBackendWebpkiCount.toInt(),
+  defaultBackendCustomCount: s.defaultBackendCustomCount.toInt(),
   androidPlatformInitSucceeded: s.androidPlatformInitSucceeded,
-  androidHybridFallbackCount: s.androidHybridFallbackCount,
+  androidHybridFallbackCount: s.androidHybridFallbackCount.toInt(),
 );
